some changes, added exception handler, deploy docker

+++ src/main/java/com/example/bankcards/service/impl/CardServiceImpl.java
@@ -1,23 +1,25 @@
 package com.example.bankcards.service.impl;
 
 import com.example.bankcards.dto.CardDTO;
-import com.example.bankcards.payload.CreateCardRequestDTO;
+import com.example.bankcards.exception.NotFoundException;
+import com.example.bankcards.exception.ValidationException;
+import com.example.bankcards.payload.request.CreateCardRequest;
 import com.example.bankcards.entity.Card;
 import com.example.bankcards.entity.User;
 import com.example.bankcards.entity.enums.CardStatus;
 import com.example.bankcards.mapper.CardMapper;
+import com.example.bankcards.payload.response.CardNumberResponse;
 import com.example.bankcards.repository.CardRepository;
 import com.example.bankcards.repository.UserRepository;
 import com.example.bankcards.service.CardService;
+import com.example.bankcards.util.Util;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 
 import java.math.BigDecimal;
 import java.time.LocalDate;
-import java.util.Base64;
 import java.util.List;
-import java.util.UUID;
 
 @Service
 @RequiredArgsConstructor
@@ -26,23 +28,32 @@ public class CardServiceImpl implements CardService {
     private final UserRepository userRepo;
 
     @Override
-    public CardDTO createCard(CreateCardRequestDTO dto) {
+    public CardDTO createCard(CreateCardRequest dto) {
         User user = userRepo.findById(dto.getUserId())
-                .orElseThrow(() -> new RuntimeException("User not found"));
+                .orElseThrow(() -> new NotFoundException("Пользователь не найден"));
 
-        Card card = new Card();
-        card.setOwnerName(dto.getOwnerName());
-        card.setExpiryDate(LocalDate.now().plusYears(2));
-        card.setBalance(BigDecimal.ZERO);
-        card.setUser(user);
-        card.setStatus(CardStatus.ACTIVE);
+        Card card = Card.builder()
+                .ownerName(dto.getOwnerName())
+                .expiryDate(LocalDate.now().plusYears(2))
+                .balance(BigDecimal.ZERO)
+                .user(user)
+                .status(CardStatus.ACTIVE)
+                .build();
 
-        String cardNumber = generateCardNumber();
-        String encrypted = encrypt(cardNumber);
-        card.setEncryptedCardNumber(cardNumber);
+        String cardNumber = Util.generateCardNumber();
+        String encrypted = Util.encrypt(cardNumber);
+        card.setEncryptedCardNumber(encrypted);
 
         Card saved = cardRepo.save(card);
         return CardMapper.toDto(saved);
+    }
+
+    @Override
+    public CardNumberResponse getFullCardNumber(Long id) {
+        Card card = cardRepo.findById(id).orElseThrow(() -> new NotFoundException("Карта не найдена"));
+        CardNumberResponse responseDTO = new CardNumberResponse();
+        responseDTO.setCardNumber(Util.decrypt(card.getEncryptedCardNumber()));
+        return responseDTO;
     }
 
     @Override
@@ -54,17 +65,21 @@ public class CardServiceImpl implements CardService {
 
     @Override
     public void requestBlock(Long cardId, Long userId) {
-        Card card = cardRepo.findById(cardId).orElseThrow();
-        if (!card.getUser().getId().equals(userId)) throw new AccessDeniedException("Forbidden");
+        Card card = cardRepo.findById(cardId).orElseThrow(() -> new NotFoundException("Карта не найдена"));
+        if (!card.getUser().getId().equals(userId)) {
+            throw new AccessDeniedException("Доступ запрещен");
+        }
         card.setStatus(CardStatus.BLOCKED);
         cardRepo.save(card);
     }
 
     @Override
     public void transfer(Long fromId, Long toId, BigDecimal amount) {
-        Card from = cardRepo.findById(fromId).orElseThrow();
-        Card to = cardRepo.findById(toId).orElseThrow();
-        if (from.getBalance().compareTo(amount) < 0) throw new RuntimeException("Insufficient balance");
+        Card from = cardRepo.findById(fromId).orElseThrow(() -> new NotFoundException("Карта отправителя не найдена"));
+        Card to = cardRepo.findById(toId).orElseThrow(() -> new NotFoundException("Карта получателя не найдена"));
+        if (from.getBalance().compareTo(amount) < 0) {
+            throw new ValidationException("Недостаточно средств");
+        }
         from.setBalance(from.getBalance().subtract(amount));
         to.setBalance(to.getBalance().add(amount));
         cardRepo.save(from);
@@ -73,13 +88,16 @@ public class CardServiceImpl implements CardService {
 
     @Override
     public void blockByAdmin(Long cardId) {
-        Card card = cardRepo.findById(cardId).orElseThrow();
+        Card card = cardRepo.findById(cardId).orElseThrow(() -> new NotFoundException("Карта не найдена"));
         card.setStatus(CardStatus.BLOCKED);
         cardRepo.save(card);
     }
 
     @Override
     public void deleteCard(Long cardId) {
+        if (!cardRepo.existsById(cardId)) {
+            throw new NotFoundException("Карта не найдена");
+        }
         cardRepo.deleteById(cardId);
     }
 
@@ -90,16 +108,6 @@ public class CardServiceImpl implements CardService {
                 .toList();
     }
 
-    private String generateCardNumber() {
-        return UUID.randomUUID().toString().replaceAll("[^0-9]", "").substring(0, 16);
-    }
 
-    private String encrypt(String number) {
-        return Base64.getEncoder().encodeToString(number.getBytes());
-    }
 
-    private String mask(String encrypted) {
-        String decoded = new String(Base64.getDecoder().decode(encrypted));
-        return "**** **** **** " + decoded.substring(12);
-    }
 }
